+++ aws_quota/check/appmesh.py
@@ -3,7 +3,7 @@ from .quota_check import QuotaCheck, QuotaScope
 
 class MeshCountCheck(QuotaCheck):
     key = "am_mesh_count"
-    description = "App Meshes per account"
+    description = "App Meshes per Account"
     scope = QuotaScope.ACCOUNT
     service_code = 'appmesh'
     quota_code = 'L-AC861A39'