+++ aws_quota/check/autoscaling.py
@@ -3,7 +3,7 @@ from .quota_check import QuotaCheck, QuotaScope
 
 class AutoScalingGroupCountCheck(QuotaCheck):
     key = "asg_count"
-    description = "Auto Scaling groups per region"
+    description = "Auto Scaling groups per Region"
     scope = QuotaScope.REGION
     service_code = 'autoscaling'
     quota_code = 'L-CDE20ADC'
@@ -15,7 +15,7 @@ class AutoScalingGroupCountCheck(QuotaCheck):
 
 class LaunchConfigurationCountCheck(QuotaCheck):
     key = "lc_count"
-    description = "Launch configurations per region"
+    description = "Launch configurations per Region"
     scope = QuotaScope.REGION
     service_code = 'autoscaling'
     quota_code = 'L-6B80B8FA'