+++ aws_quota/check/dynamodb.py
@@ -3,7 +3,7 @@ from .quota_check import QuotaCheck, QuotaScope
 
 class TableCountCheck(QuotaCheck):
     key = "dyndb_table_count"
-    description = "DynamoDB Tables per region"
+    description = "DynamoDB Tables per Region"
     scope = QuotaScope.REGION
     service_code = 'dynamodb'
     quota_code = 'L-F98FE922'