+++ aws_quota/check/ebs.py
@@ -3,7 +3,7 @@ from .quota_check import QuotaCheck, QuotaScope
 
 class SnapshotCountCheck(QuotaCheck):
     key = "ebs_snapshot_count"
-    description = "EBS Snapshots per region"
+    description = "EBS Snapshots per Region"
     scope = QuotaScope.REGION
     service_code = 'ebs'
     quota_code = 'L-309BACF6'