+++ aws_quota/check/ecs.py
@@ -3,7 +3,7 @@ from .quota_check import QuotaCheck, QuotaScope
 
 class ClusterCountCheck(QuotaCheck):
     key = "ecs_count"
-    description = "ECS Clusters per region"
+    description = "ECS Clusters per Region"
     scope = QuotaScope.REGION
     service_code = 'ecs'
     quota_code = 'L-21C621EB'