+++ aws_quota/check/route53resolver.py
@@ -3,7 +3,7 @@ from .quota_check import QuotaCheck, QuotaScope
 
 class EndpointCountCheck(QuotaCheck):
     key = "route53resolver_endpoint_count"
-    description = "Route53 Resolver endpoints per region"
+    description = "Route53 Resolver endpoints per Region"
     scope = QuotaScope.REGION
     service_code = 'route53resolver'
     quota_code = 'L-4A669CC0'
@@ -14,7 +14,7 @@ class EndpointCountCheck(QuotaCheck):
 
 class RulesCountCheck(QuotaCheck):
     key = "route53resolver_rule_count"
-    description = "Route53 Resolver rules per region"
+    description = "Route53 Resolver rules per Region"
     scope = QuotaScope.REGION
     service_code = 'route53resolver'
     quota_code = 'L-51D8A1FB'
@@ -25,7 +25,7 @@ class RulesCountCheck(QuotaCheck):
 
 class RuleAssociationsCountCheck(QuotaCheck):
     key = "route53resolver_rule_association_count"
-    description = "Route53 Resolver rule associations per region"
+    description = "Route53 Resolver rule associations per Region"
     scope = QuotaScope.REGION
     service_code = 'route53resolver'
     quota_code = 'L-94E19253'