+++ aws_quota/check/s3.py
@@ -3,7 +3,7 @@ from .quota_check import QuotaCheck, QuotaScope
 
 class BucketCountCheck(QuotaCheck):
     key = "s3_bucket_count"
-    description = "S3 Buckets per region"
+    description = "S3 Buckets per Account"
     scope = QuotaScope.REGION
     service_code = 's3'
     quota_code = 'L-DC2B2D3D'