+++ aws_quota/check/secretsmanager.py
@@ -3,7 +3,7 @@ from .quota_check import QuotaCheck, QuotaScope
 
 class SecretCountCheck(QuotaCheck):
     key = "secretsmanager_secrets_count"
-    description = "Secrets per account"
+    description = "Secrets per Account"
     scope = QuotaScope.ACCOUNT
     service_code = 'secretsmanager'
     quota_code = 'L-2F66C23C'