+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - add error handling for CLI result reporter
+- new check: iam_role_count
 
 ## [1.9.0] - 2021-09-21